% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{topTable,pamClass-method}
\alias{topTable,pamClass-method}
\title{Top table for \code{pamClass} object}
\usage{
\S4method{topTable}{pamClass}(fit, n)
}
\arguments{
\item{fit}{object for which to obtain a top table, generally a fit object for a given model class}

\item{n}{number of features (variables) to list in the top table, ranked by importance}
}
\value{
\code{topTablePam} object
}
\description{
Top table for \code{pamClass} object
}
