% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{topTable,rfClass-method}
\alias{topTable,rfClass-method}
\title{Top table for \code{rfClass} object}
\usage{
\S4method{topTable}{rfClass}(fit, n)
}
\arguments{
\item{fit}{object for which to obtain a top table, generally a fit object for a given model class}

\item{n}{number of features (variables) to list in the top table, ranked by importance}
}
\value{
\code{topTableRfClass} object
}
\description{
Top table for \code{rfClass} object
}
